third_party/blink/web_tests/printing/page-break-after-avoid.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/paged-media-test-utils.js"></script>
<style>
  #sandbox > * { break-inside: avoid; }
</style>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="sandbox"></div>
<script src="script-tests/page-break-after-avoid.js"></script>
<script>runPrintingTest(test);</script>
</body>
</html>
