third_party/blink/web_tests/printing/page-break-after-avoid-expected.txt

Test for page-break-after:avoid

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "page1" is 1
PASS: page number of "page2-1" is 2
PASS: page number of "page2-2" is 2
PASS: page number of "page3-1" is 3
PASS: page number of "page3-2" is 3
PASS: page number of "page3-3" is 3
PASS: page number of "page5" is 5
PASS: page number of "page6-1" is 6
PASS: page number of "page6-7" is 6
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
